aiscale/requirements.txt
aenum                  3.1.16
annotated-types        0.7.0
bcrypt                 5.0.0
bidict                 0.23.1
blinker                1.9.0
cachelib               0.9.0
certifi                2025.8.3
charset-normalizer     3.4.3
click                  8.1.8
colorama               0.4.6
contourpy              1.3.3
cycler                 0.12.1
dnspython              2.7.0
dominate               2.9.1
edgeimpulse            1.0.18
edgeimpulse-api        1.74.21
email_validator        2.2.0
filelock               3.20.0
Flask                  3.1.0
Flask-Admin            1.6.1
Flask-Bootstrap        3.3.7.1
Flask-Caching          2.3.0
Flask-Cors             5.0.0
Flask-Login            0.6.3
Flask-Principal        0.4.0
Flask-Security         5.5.2
Flask-SQLAlchemy       3.1.1
Flask-WTF              1.2.2
fonttools              4.60.1
fsspec                 2025.9.0
greenlet               3.1.1
gunicorn               23.0.0
h11                    0.16.0
idna                   3.10
importlib_resources    6.5.2
itsdangerous           2.2.0
Jinja2                 3.1.5
kiwisolver             1.4.9
MarkupSafe             3.0.2
matplotlib             3.10.7
mpmath                 1.3.0
networkx               3.5
numpy                  2.2.6
opencv-python          4.12.0.88
opencv-python-headless 4.12.0.88
packaging              25.0
passlib                1.7.4
pillow                 12.0.0
pip                    25.3
polars                 1.34.0
polars-runtime-32      1.34.0
psutil                 7.1.0
psycopg2-binary        2.9.11
pydantic               2.11.9
pydantic_core          2.33.2
PyJWT                  2.10.1
pyparsing              3.2.5
pyserial               3.5
python-dateutil        2.9.0.post0
python-engineio        4.12.2
python-socketio        5.13.0
PyYAML                 6.0.3
requests               2.32.5
scipy                  1.16.2
setuptools             65.5.0
simple-websocket       1.1.0
six                    1.17.0
SQLAlchemy             2.0.37
sympy                  1.14.0
torch                  2.9.0
torchvision            0.24.0
typing_extensions      4.12.2
typing-inspection      0.4.1
tzdata                 2025.2
ultralytics            8.3.217
ultralytics-thop       2.0.17
urllib3                1.26.20
visitor                0.1.3
websocket-client       1.8.0
Werkzeug               3.1.3
wsproto                1.2.0
WTForms                3.2.1